requirements.txt
pdfminer.six
pytest
python-dotenv
requests
